+++ src/GRA.Data.SqlServer/Migrations/20251014221117_add-achiever-report-criterion.Designer.cs
@@ -4,6 +4,7 @@ using GRA.Data.SqlServer;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -11,9 +12,11 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace GRA.Data.SqlServer.Migrations
 {
     [DbContext(typeof(SqlServerContext))]
-    partial class SqlServerContextModelSnapshot : ModelSnapshot
+    [Migration("20251014221117_add-achiever-report-criterion")]
+    partial class addachieverreportcriterion
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -256,26 +259,6 @@ namespace GRA.Data.SqlServer.Migrations
                     b.ToTable("AvatarColors");
                 });
 
-            modelBuilder.Entity("GRA.Data.Model.AvatarColorText", b =>
-                {
-                    b.Property<int>("AvatarColorId")
-                        .HasColumnType("int");
-
-                    b.Property<int>("LanguageId")
-                        .HasColumnType("int");
-
-                    b.Property<string>("AltText")
-                        .IsRequired()
-                        .HasMaxLength(130)
-                        .HasColumnType("nvarchar(130)");
-
-                    b.HasKey("AvatarColorId", "LanguageId");
-
-                    b.HasIndex("LanguageId");
-
-                    b.ToTable("AvatarColorTexts");
-                });
-
             modelBuilder.Entity("GRA.Data.Model.AvatarElement", b =>
                 {
                     b.Property<int>("Id")
@@ -347,26 +330,6 @@ namespace GRA.Data.SqlServer.Migrations
                     b.HasIndex("AvatarLayerId");
 
                     b.ToTable("AvatarItems");
-                });
-
-            modelBuilder.Entity("GRA.Data.Model.AvatarItemText", b =>
-                {
-                    b.Property<int>("AvatarItemId")
-                        .HasColumnType("int");
-
-                    b.Property<int>("LanguageId")
-                        .HasColumnType("int");
-
-                    b.Property<string>("AltText")
-                        .IsRequired()
-                        .HasMaxLength(130)
-                        .HasColumnType("nvarchar(130)");
-
-                    b.HasKey("AvatarItemId", "LanguageId");
-
-                    b.HasIndex("LanguageId");
-
-                    b.ToTable("AvatarItemTexts");
                 });
 
             modelBuilder.Entity("GRA.Data.Model.AvatarLayer", b =>
@@ -2631,11 +2594,6 @@ namespace GRA.Data.SqlServer.Migrations
                         .HasMaxLength(500)
                         .HasColumnType("nvarchar(500)");
 
-                    b.Property<string>("ContactName")
-                        .IsRequired()
-                        .HasMaxLength(255)
-                        .HasColumnType("nvarchar(255)");
-
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
 
@@ -2669,10 +2627,9 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Property<bool>("PhonePreferred")
                         .HasColumnType("bit");
 
-                    b.Property<string>("References")
-                        .IsRequired()
-                        .HasMaxLength(2000)
-                        .HasColumnType("nvarchar(2000)");
+                    b.Property<string>("ReferencesFilename")
+                        .HasMaxLength(255)
+                        .HasColumnType("nvarchar(255)");
 
                     b.Property<bool>("RegistrationCompleted")
                         .HasColumnType("bit");
@@ -2809,9 +2766,6 @@ namespace GRA.Data.SqlServer.Migrations
                         .HasMaxLength(375)
                         .HasColumnType("nvarchar(375)");
 
-                    b.Property<bool>("IsApproved")
-                        .HasColumnType("bit");
-
                     b.Property<int>("MaximumCapacity")
                         .HasColumnType("int");
 
@@ -2900,14 +2854,6 @@ namespace GRA.Data.SqlServer.Migrations
                         .HasColumnType("nvarchar(255)");
 
                     b.Property<string>("ContactEmail")
-                        .HasMaxLength(255)
-                        .HasColumnType("nvarchar(255)");
-
-                    b.Property<string>("CoverSheetBranch")
-                        .HasMaxLength(255)
-                        .HasColumnType("nvarchar(255)");
-
-                    b.Property<string>("CoverSheetContact")
                         .HasMaxLength(255)
                         .HasColumnType("nvarchar(255)");
 
@@ -4613,25 +4559,6 @@ namespace GRA.Data.SqlServer.Migrations
                     b.Navigation("AvatarLayer");
                 });
 
-            modelBuilder.Entity("GRA.Data.Model.AvatarColorText", b =>
-                {
-                    b.HasOne("GRA.Data.Model.AvatarColor", "AvatarColor")
-                        .WithMany("Texts")
-                        .HasForeignKey("AvatarColorId")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.HasOne("GRA.Data.Model.Language", "Language")
-                        .WithMany()
-                        .HasForeignKey("LanguageId")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("AvatarColor");
-
-                    b.Navigation("Language");
-                });
-
             modelBuilder.Entity("GRA.Data.Model.AvatarElement", b =>
                 {
                     b.HasOne("GRA.Data.Model.AvatarColor", "AvatarColor")
@@ -4659,25 +4586,6 @@ namespace GRA.Data.SqlServer.Migrations
                         .IsRequired();
 
                     b.Navigation("AvatarLayer");
-                });
-
-            modelBuilder.Entity("GRA.Data.Model.AvatarItemText", b =>
-                {
-                    b.HasOne("GRA.Data.Model.AvatarItem", "AvatarItem")
-                        .WithMany("Texts")
-                        .HasForeignKey("AvatarItemId")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.HasOne("GRA.Data.Model.Language", "Language")
-                        .WithMany()
-                        .HasForeignKey("LanguageId")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("AvatarItem");
-
-                    b.Navigation("Language");
                 });
 
             modelBuilder.Entity("GRA.Data.Model.AvatarLayerText", b =>
@@ -5717,16 +5625,6 @@ namespace GRA.Data.SqlServer.Migrations
             modelBuilder.Entity("GRA.Data.Model.AvatarBundle", b =>
                 {
                     b.Navigation("AvatarBundleItems");
-                });
-
-            modelBuilder.Entity("GRA.Data.Model.AvatarColor", b =>
-                {
-                    b.Navigation("Texts");
-                });
-
-            modelBuilder.Entity("GRA.Data.Model.AvatarItem", b =>
-                {
-                    b.Navigation("Texts");
                 });
 
             modelBuilder.Entity("GRA.Data.Model.AvatarLayer", b =>