+++ src/GRA.Domain.Repository/IUserLogRepository.cs
@@ -10,9 +10,16 @@ namespace GRA.Domain.Repository
     {
         Task<Dictionary<string, long>> ActivityEarningsTotalAsync(ReportCriterion request);
 
-        Task<long> CompletedChallengeCountAsync(ReportCriterion request, int? challengeId = null);
+        Task<(long earned, long earnedAchiever)>
+            CompletedChallengeCountAsync(ReportCriterion request);
 
-        Task<long> EarnedBadgeCountAsync(ReportCriterion request, int? badgeId = null);
+        Task<(long earned, long earnedAchiever)>
+            CompletedChallengeCountAsync(ReportCriterion request, int? challengeId);
+
+        Task<(long earned, long earnedAchiever)> EarnedBadgeCountAsync(ReportCriterion request);
+
+        Task<(long earned, long earnedAchiever)>
+            EarnedBadgeCountAsync(ReportCriterion request, int? badgeId);
 
         Task<int> GetActivityEarnedForUserAsync(int userId);